requirements.txt
accumulation-tree==0.6.2
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
cloudpickle==2.2.1
filelock==3.12.3
frozenlist==1.4.0
fsspec==2023.6.0
gym==0.26.2
gym-notices==0.0.8
idna==3.4
Jinja2==3.1.2
lightning-utilities==0.9.0
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.25.2
packaging==23.1
pytorch-lightning==2.0.7
pyudorandom==1.0.0
PyYAML==6.0.1
requests==2.31.0
sympy==1.12
tdigest==0.5.2.2
torch==2.0.1
torchmetrics==1.1.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2

